requirements.txt
numpy==1.26.2
scipy==1.11.4
torch==2.1.2
accelerate==0.25.0
bitsandbytes==0.41.3.post2
peft==0.7.1
transformers==4.35.2
ctransformers[cuda]>=0.2.24
pillow==10.1.0
llama-cpp-python==0.2.23
datasets
zstandard
jsonlines
wandb
openai==1.3.9
sentencepiece==0.1.99
fire
datasketch==1.6.4
nltk==3.8.1
scikit-learn==1.3.2
pytest
wheel
flash-attn==2.3.6
llm-rs==0.2.15
